+++ lib/slimcop.rb
@@ -5,6 +5,7 @@ require_relative 'slimcop/version'
 module Slimcop
   autoload :Cli, 'slimcop/cli'
   autoload :Configuration, 'slimcop/configuration'
+  autoload :FileCollector, 'slimcop/file_collector'
   autoload :Formatter, 'slimcop/formatter'
   autoload :Offense, 'slimcop/offense'
   autoload :RubyExtractor, 'slimcop/ruby_extractor'